config: add --config flag and tests
Added flag so we can now specify the path to a config file (used to
override the default walker.yaml, which will also be read if it exists).
Noticed in testing that gopkg.in/yaml.v1 does not error on type mismatch
but v2 does, so upgraded to that.

You may need to get rid of `src/gopkg.in/yaml.v1` and/or run `go clean`
and `go get` again to make everything work.