fetcher: when transport is nil, set a default one
This is so that we can override the DNS cache. It would be great to
confirm that it's actually using the DNS cache as well (though we do log
it if it isn't), but it seems like replacing the Dial function doesn't
afford us a way to check that. At some point we may need a bit more
abstraction around the Transport to verify it.