Adding coverage script.
NOTE: this command will break on the command line. The problem is that
somewhere in the (go) testing code a chdir is being done. That means
that our relative path to TemplatesDirectory is no long good, and
so console.Start aborts. I don't know what Dan wants as a fix.