fetcher: do not exclude entire hosts for TCP error
Previously we assumed we shouldn't crawl a host if we couldn't connect
to it, but in some cases (ex. rackcdn.com) the subdomains could be
crawled. In the future we should do more to check the IP on a per-fetch
basis, but for now changed it so we still try to crawl the site.